+++ foxes/data/__init__.py
@@ -1,2 +1,2 @@
-from .parse import parse_Pct_file_name
+from .parse import parse_Pct_file_name, parse_Pct_two_files
 from .static_data import FARM, STATES, PCTCURVE, StaticData