+++ foxes/models/point_models/__init__.py
@@ -1,1 +1,2 @@
 from .wake_deltas import WakeDeltas
+from .set_uniform_data import SetUniformData