+++ docs/source/history.md
@@ -75,7 +75,7 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 - Input:
   - New input: `MultiHeightStates` and `MultiHeightTimeseries`, for input data at several height levels, e.g. WRF results at one point
 - Output:
-  - New output: `FarmResultsEval`, calculates sum, mean, min, max over states or turbines for the whole wind farm
+  - New output: `FarmResultsEval`, calculates sum, mean, min, max over states or turbines for the whole wind farm. Also calculates capacity, efficiency, yield, P75, P90.
   - New output: `ResultsWriter`, writes farm results or pandas data to csv file
   - Renaming: `AmbientRosePlotOutput` is now called `StatesRosePlotOutput`
 - Notebooks:
@@ -119,3 +119,25 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
   - Hotfix for bug in `TurbineTypeCurves` output
 
 **Full Changelog**: [](https://github.com/FraunhoferIWES/foxes/commits/v0.2.4)
+
+## v0.2.5-alpha
+
+- Core:
+  - New: `ExtendedStates`, allows for adding point data models to states
+- Input:
+  - Improved: `FieldDataNC`, now with support for spatially uniform variables
+- Turbine types:
+  - New: `WsRho2PCtTwoFiles`, air density dependent power and ct curves
+- Turbine models:
+  - New: `SectorManagement`, sets variables by range rules on other variables
+- Point models:
+  - New: `SetUniformData`, set uniform variables (optionally state dependent)
+- Examples:
+  - New: `sector_management`, demonstrates how to model wind sector management
+- Notebooks:
+  - New: `sector_man.ipynb`, demonstrates how to model wind sector management
+  - New: `data.ipynb`, lists and shows the static data
+- Bug fixes:
+  - Fix for bug with option `col_models` in farm layout from csv
+
+**Full Changelog**: [](https://github.com/FraunhoferIWES/foxes/commits/v0.2.5)